chore: clear 75 lint issues + enterprise-tone rewrite of README/SECURITY/CONTRIBUTING
CI-lint bucket:
- 50 errcheck sites: cleanup-path Close/Kill/RemoveAll get //nolint:errcheck
  with justifications; meaningful ExecContext / enc.Encode errors are logged
  or returned; stdout Fprintf in CLI keeps //nolint (unrecoverable, rare).
- 6 unused symbols deleted (discardReader, streamDelta, streamingStubRunner).
- 3 unconvert conversions removed (openai, matrix).
- 2 bodyclose sites (discord/slack ws handshake) now defer resp.Body.Close().
- 2 nilerr WalkDir callbacks documented (intentional continue-on-error).
- 2 staticcheck: SA1019 CredentialsFromJSON -> CredentialsFromJSONWithParams;
  SA9003 empty branch removed.
- 1 copyloopvar, 1 ineffassign, 1 errorlint, 1 gocritic (ifElseChain -> switch),
  1 revive (WhatsAppConfig godoc), 4 forbidigo (Must*/example) with justification.

Docs bucket:
- README: reframed from "written by one engineer" to enterprise positioning.
  Lists all 9 transports, 5 providers, MCP, cron, skills, compression, recall,
  SLSA-3, cosign, SBOM. Adds Kubernetes/OpenShift note, cosign verify example,
  full config schema (bedrock/vertex/approver/all transports).
- SECURITY.md: response SLA table by severity, cryptography inventory,
  coordinated-disclosure guidance, expanded supply-chain matrix, cosign
  verification recipe.
- CONTRIBUTING.md: reviewer checklist, release process, governance section,
  explicit test-injection guidance.

Everything green: build ok, race-tests pass, golangci-lint 0 issues,
coverage 76.0%.



---
THE ARCHITECT ᛫ Sebastien Rousseau ᛫ https://sebastienrousseau.com
THE ENGINE ᛞ EUXIS ᛫ Enterprise Unified Execution Intelligence System ᛫ https://euxis.co

+++ internal/transport/matrix/client.go
@@ -135,7 +135,7 @@ func (c *Client) sync(ctx context.Context) (*syncResponse, error) {
 	c.mu.Unlock()
 
 	q := url.Values{}
-	q.Set("timeout", strconv.FormatInt(int64(c.cfg.PollTimeout.Milliseconds()), 10))
+	q.Set("timeout", strconv.FormatInt(c.cfg.PollTimeout.Milliseconds(), 10))
 	if since != "" {
 		q.Set("since", since)
 	}
